Introduce separate statements page (#922)
* Remove duplicate color vars

* Custom css for select elements

Use a custom appearance for `<select>` elements, including a light gray border and drop-down arrow.

* Split “Add Channel” button and modal into its own partial

* Add xlarge panel size

* Split statements page from dashboard
