+++ app/assets/stylesheets/_colors.scss
@@ -47,13 +47,6 @@ $braveGray-8: #dbdfe3;
 $braveGray-9: #e7ebee;
 $braveGray-10: #f7f8f9;
 
-$braveInfo-Background: #eefbfd;
-$braveInfo-Foreground: #00bcd6;
-$braveInfo-Border: #e7ebee;
-
-$braveError-Background: #feedf0;
-$braveError-Foreground: #dd0529;
-
 $braveWarning-Background: #FFF8D2;
 $braveWarning-Foreground: $braveGray-1;
