+++ app/assets/stylesheets/pages/statements.scss
@@ -1,0 +1,104 @@
+#statements_panel {
+  .single-panel--padded-content {
+    padding: 4rem 6rem;
+
+    @include media-breakpoint-down(sm) {
+      padding: 3rem 1.5rem;
+    }
+  }
+
+
+  #statement_generator {
+    margin-top: 2.5rem;
+
+    .btn-primary {
+      background-color: $braveAction;
+      border: none;
+      margin-left: 0.8rem;
+      &:hover, &:active, &:focus {
+        border: none;
+        box-shadow: none;
+      }
+      &:hover {
+        background-color: lighten($braveAction, 5%);
+      }
+    }
+
+    @include media-breakpoint-down(sm) {
+      select {
+        width: 100%;
+      }
+      .btn-primary {
+        margin: 1.5rem 0 0 0;
+        width: 100%;
+      }
+    }
+  }
+
+  .statements-table {
+    margin-top: 2.5rem;
+
+    .content {
+      min-height: 150px;
+
+      &:empty {
+        &:after {
+          content: "You haven't requested a statement yet!";
+          display: block;
+          padding-top: 130px;
+          text-align: center;
+          color: $braveGray-7;
+          font-size: 1.2rem;
+        }
+        background-image: url(asset-path("icn-blankslate-statement.svg"));
+        background-repeat: no-repeat;
+        background-position: top 40px center;
+      }
+    }
+
+    .statement, .header {
+      font-size: 14px;
+      font-weight: 300;
+      padding: 0.6rem 0;
+
+      .created-at, .period, .status {
+        display: inline-block;
+        text-overflow: ellipsis;
+        white-space: nowrap;
+      }
+      .created-at {
+        width: 28%;
+      }
+      .period {
+        width: 32%;
+      }
+      .status {
+        width: 40%;
+      }
+    }
+
+    .header {
+      color: $braveInfo-Foreground;
+      font-size: 12px;
+      font-weight: 600;
+      text-transform: uppercase;
+      border-bottom: 2px solid $lightGray;
+    }
+
+    .statement {
+      font-weight: 300;
+      border-bottom: 1px solid $lightGray;
+      .status {
+        .download {
+          float: right;
+          background-image: url(asset-path("download@1x.png"));
+          background-position: center left;
+          background-repeat: no-repeat;
+          background-size: 17px 21px;
+          padding-left: 25px;
+        }
+      }
+    }
+  }
+
+}
