+++ app/javascript/packs/application.js
@@ -6,6 +6,7 @@ import 'utils/request';
 import 'u2f/registrationPage';
 import 'u2f/authenticationPage';
 import 'publishers/home';
+import 'publishers/statements';
 import Rails from 'rails-ujs';
 
 /*
