+++ app/views/publishers/statements.html.slim
@@ -1,0 +1,39 @@
+noscript
+  div.noscript-warning = t ".noscript"
+
+- content_for(:navbar_content_right) do
+  = link_to( \
+    t("shared.dashboard"), \
+    home_publishers_path, \
+    class: 'title' \
+  )
+  = render partial: "choose_channel_button"
+
+.single-panel--wrapper.single-panel--wrapper--xlarge#statements_panel
+  = render "panel_flash_messages"
+  .single-panel--content
+    .single-panel--padded-content.text-left
+      h3= t ".title"
+      p= t ".description"
+      - if unused_statement_periods.length > 0
+        = form_for(current_publisher, url: generate_statement_publishers_path, html: { id: "statement_generator" }) do |f|
+          .form-group
+            = select_tag(:statement_period, options_for_select(statement_periods_as_options(unused_statement_periods), :past_30_days))
+            a.btn.btn-primary#generate_statement href="#"
+              = t ".create_statement"
+      .statements-table
+        .header
+          .created-at= t ".created_at"
+          .period= t ".period"
+          .status= t ".status"
+        .content#generated_statements
+          - current_publisher.statements.each do |s|
+            .statement
+              .created-at= statement_period_date(s.created_at)
+              .period= statement_period_description(s.period.to_sym)
+              .status
+                - if s.encrypted_contents?
+                  span= 'Ready'
+                  = link_to(t("shared.download"), statement_publishers_url(id: s.id), class: 'download')
+                - else
+                  = t ".statements.delayed"
