+++ test/controllers/publishers_controller_test.rb
@@ -579,8 +579,8 @@ class PublishersControllerTest < ActionDispatch::IntegrationTest
     assert_response 200
     assert_match(
       '{"id":"' + publisher_statement.id + '",' +
-        '"date":"' + publisher_statement.created_at.strftime('%B %e, %Y') + '",' +
-        '"period":"All"}',
+        '"date":"' + publisher_statement.created_at.strftime('%B %e') + '",' +
+        '"period":"All dates"}',
       response.body)
   end
